refactor: split ChatView.swift into focused files
Extract ChatBubble, markdown parser, and widget views from the 2,285-line
ChatView.swift into dedicated files. No logic changes — only code movement
and access modifier adjustments (private → internal for cross-file visibility).
